2,3,4 exercise

+++ OrderFlow/Models/Product.cs
@@ -3,7 +3,7 @@ public class Product
     public double Price {get; private set;}
     public int Stock {get; private set;}
     public string Category {get; private set;}
-    public Product (int price, int stock, string category)
+    public Product (double price, int stock, string category)
     {
         Price = price;
         Stock = stock;
